🚚(backend) move assertLogsContains to BaseLogMixinTestCase
A convenient assertLogsContains method was existing in a specialized
TestCase class. Is has been moved to BaseLogMixinTestCase for ease of
use.
