+++ src/backend/joanie/tests/edx_imports/base_test_commands_migrate.py
@@ -3,6 +3,7 @@
 from django.test import TestCase, override_settings
 
 from joanie.edx_imports.edx_factories import session
+from joanie.tests.base import BaseLogMixinTestCase
 
 
 @override_settings(
@@ -27,7 +28,7 @@ from joanie.edx_imports.edx_factories import session
     EDX_TIME_ZONE="UTC",
     TIME_ZONE="UTC",
 )
-class MigrateOpenEdxBaseTestCase(TestCase):
+class MigrateOpenEdxBaseTestCase(TestCase, BaseLogMixinTestCase):
     """Base test case for the migrate command."""
 
     maxDiff = None
@@ -35,20 +36,3 @@ class MigrateOpenEdxBaseTestCase(TestCase):
     def tearDown(self):
         """Tear down the test case."""
         session.rollback()
-
-    def assertLogsContains(self, logger, expected_records):
-        """
-        Assert that the logger contains the expected messages and levels.
-        """
-        records = [record.getMessage() for record in logger.records]
-        for expected_record in expected_records:
-            is_found = False
-            for record in records:
-                try:
-                    self.assertIn(expected_record, record)
-                    is_found = True
-                    break
-                except AssertionError:
-                    pass
-            if not is_found:
-                self.fail(f"Expected record {expected_record} not found in {records}")
